fix: move integration tests into integration directory
