+++ frontend/src/test/integration/routes/bottles/new.test.tsx
@@ -2,8 +2,8 @@ import { describe, it, expect, vi, beforeEach } from 'vite-plus/test';
 import { screen, waitFor, within } from '@testing-library/react';
 import { createMockAuthState, renderWithFileRoutes } from '../../file-route-utils';
 import userEvent from '@testing-library/user-event';
-import type { BottleResponseModel } from '../../../api/generated/bottles-api';
-import { customBottlesInstance } from '../../../api/axios-instance';
+import type { BottleResponseModel } from '../../../../api/generated/bottles-api';
+import { customBottlesInstance } from '../../../../api/axios-instance';
 
 function renderNewBottleRouteWithAuth() {
   return renderWithFileRoutes({
@@ -102,8 +102,8 @@ function returnBottleResponse(): BottleResponseModel {
   };
 }
 
-vi.mock('../../../api/axios-instance', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('../../../api/axios-instance')>();
+vi.mock('../../../../api/axios-instance', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('../../../../api/axios-instance')>();
   const bottleResponse = returnBottleResponse();
 
   return {
